Updated some `NONE` usages

+++ game/source/game/game_engine.cpp
@@ -269,7 +269,7 @@ void __cdecl game_engine_interface_update(real world_seconds_elapsed)
 					c_gui_screen_scoreboard* screen = c_gui_screen_scoreboard::get_scoreboard_screen(controller_index);
 					if (!screen || !window_manager_get()->get_screen_above(screen->get_render_window(), screen))
 					{
-						bool show_scoreboard = back_pressed && (game_is_cooperative() || game_is_multiplayer()) && player_control_get_zoom_level(user_index) == (short)NONE;
+						bool show_scoreboard = back_pressed && (game_is_cooperative() || game_is_multiplayer()) && player_control_get_zoom_level(user_index) == NONE;
 
 						if (current_game_engine())
 						{
